Display time text in week and workWeek
