+++ js/ui/scheduler/workspaces/ui.scheduler.work_space_day.js
@@ -1,6 +1,5 @@
 import registerComponent from '../../../core/component_registrator';
 import SchedulerWorkSpaceVertical from './ui.scheduler.work_space_vertical';
-import dateLocalization from '../../../localization/date';
 
 import dxrDayDateTableLayout from '../../../renovation/ui/scheduler/workspaces/day/date_table/layout.j';
 
@@ -36,33 +35,6 @@ class SchedulerWorkSpaceDay extends SchedulerWorkSpaceVertical {
 
     _renderDateHeader() {
         return this.option('intervalCount') === 1 ? null : super._renderDateHeader();
-    }
-
-    renovatedRenderSupported() { return true; }
-
-    generateRenderOptions() {
-        const startViewDate = this._getDateWithSkippedDST();
-        const _getTimeText = (i) => {
-            // T410490: incorrectly displaying time slots on Linux
-            const index = i % this._getRowCount();
-            if(index % 2 === 0) {
-                return dateLocalization.format(this._getTimeCellDateCore(startViewDate, i), 'shorttime');
-            }
-            return '';
-        };
-
-        const options = super.generateRenderOptions();
-        options.cellDataGetters.push((_, rowIndex, cellIndex) => {
-            return {
-                value: {
-                    // TODO check the need
-                    // startDate: this._getTimeCellDate(rowIndex, cellIndex),
-                    text: _getTimeText(rowIndex, cellIndex)
-                }
-            };
-        });
-
-        return options;
     }
 
     renderRDateTable() {
