+++ js/ui/scheduler/workspaces/ui.scheduler.work_space_week.js
@@ -53,7 +53,6 @@ class SchedulerWorkSpaceWeek extends SchedulerWorkSpaceVertical {
         return super._isApplyCompactAppointmentOffset();
     }
 
-    renovatedRenderSupported() { return true; }
     renderRDateTable() {
         this.renderRComponent(
             this._$dateTable,
